access_service/pyproject.toml
[tool.poetry]
name = "messenger-access-service"
version = "0.1.0"
description = ""
authors = ["ekwize <ulanovsega@gmail.com>"]
readme = "README.md"
packages = [{include = "messenger"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.2"
dependency-injector = "^4.42.0"
sqlalchemy = "^2.0.34"
alembic = "^1.13.2"
adaptix = "^3.0.0b8"
argon2-cffi = "^23.1.0"
pytest = "^8.3.3"
uvicorn = "^0.32.0"
asyncpg = "^0.29.0"
ruff = "^0.7.0"
pyjwt = "^2.9.0"
aiokafka = "^0.11.0"
redis-async-client = "^0.1.7"
orjson = "^3.10.10"
aio-pika = "^9.5.4"
aiojobs = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
testpaths = [
    "tests",
]
pythonpath = "src"



